requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
docstring_parser==0.17.0
Flask==2.3.3
fonttools==4.60.0
gitdb==4.0.12
GitPython==3.1.37
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.117.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.39.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.6
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==2.3.7
google-auth
google-cloud-logging